Move SequenceProcessor to activity
