+++ packages/backend/src/core/activity/counters/RpcCounter.ts
@@ -8,8 +8,8 @@ import { SequenceProcessorRepository } from '../../../peripherals/database/Seque
 import { RpcClient } from '../../../peripherals/rpcclient/RpcClient'
 import { Clock } from '../../Clock'
 import { promiseAllPlus } from '../../queue/promiseAllPlus'
-import { SequenceProcessor } from '../../SequenceProcessor'
 import { RpcActivityTransactionConfig } from '../ActivityTransactionConfig'
+import { SequenceProcessor } from '../SequenceProcessor'
 import { TransactionCounter } from '../TransactionCounter'
 import { createBlockTransactionCounter } from './BlockTransactionCounter'
 import { getBatchSizeFromCallsPerMinute } from './getBatchSizeFromCallsPerMinute'
