+++ packages/backend/src/core/activity/counters/StarknetCounter.ts
@@ -8,8 +8,8 @@ import { SequenceProcessorRepository } from '../../../peripherals/database/Seque
 import { StarknetClient } from '../../../peripherals/starknet/StarknetClient'
 import { Clock } from '../../Clock'
 import { promiseAllPlus } from '../../queue/promiseAllPlus'
-import { SequenceProcessor } from '../../SequenceProcessor'
 import { SimpleActivityTransactionConfig } from '../ActivityTransactionConfig'
+import { SequenceProcessor } from '../SequenceProcessor'
 import { TransactionCounter } from '../TransactionCounter'
 import { createBlockTransactionCounter } from './BlockTransactionCounter'
 import { getBatchSizeFromCallsPerMinute } from './getBatchSizeFromCallsPerMinute'
